Move PixelThresholdClusterizer (back?) to interface+src in order to use it outside of RecoLocalTracker/SiPixelClusterizer
