+++ RecoLocalTracker/SiPixelClusterizer/src/PixelThresholdClusterizer.cc
@@ -20,8 +20,8 @@
 //----------------------------------------------------------------------------
 
 // Our own includes
-#include "PixelThresholdClusterizer.h"
-#include "SiPixelArrayBuffer.h"
+#include "RecoLocalTracker/SiPixelClusterizer/interface/PixelThresholdClusterizer.h"
+#include "RecoLocalTracker/SiPixelClusterizer/interface/SiPixelArrayBuffer.h"
 #include "CondFormats/SiPixelObjects/interface/SiPixelGainCalibrationOffline.h"
 // Geometry
 #include "Geometry/TrackerGeometryBuilder/interface/PixelGeomDetUnit.h"
